+++ src/pages/_app.tsx
@@ -11,15 +11,7 @@ import Loader from "src/components/Loader";
 import LoginDialogProvider from "../components/auth/LoginDialogProvider";
 import { DarkThemeProvider } from "../components/DarkThemeProvider";
 import { useApollo } from "../utils/apolloClient";
-import {
-    KBarProvider,
-    KBarPortal,
-    KBarPositioner,
-    KBarAnimator,
-    KBarSearch,
-    useMatches,
-    KBarResults
-} from "kbar";
+import KBar from "src/components/kbar";
 import "highlight.js/styles/default.css";
 
 export const cache = createCache({ key: "css" });
@@ -57,15 +49,7 @@ export default function MyApp({ Component, pageProps }: AppProps) {
         <CacheProvider value={cache}>
             <CssBaseline />
             <DarkThemeProvider>
-                <KBarProvider actions={actionsList}>
-                    <KBarPortal>
-                        <KBarPositioner>
-                            <KBarAnimator>
-                                <KBarSearch />
-                                <RenderResults />
-                            </KBarAnimator>
-                        </KBarPositioner>
-                    </KBarPortal>
+                <KBar>
                     <ApolloProvider client={apolloClient}>
                         <SnackbarProvider maxSnack={3}>
                             <FinishSignupDialogProvider>
@@ -79,31 +63,8 @@ export default function MyApp({ Component, pageProps }: AppProps) {
                             </FinishSignupDialogProvider>
                         </SnackbarProvider>
                     </ApolloProvider>
-                </KBarProvider>
+                </KBar>
             </DarkThemeProvider>
         </CacheProvider>
-    );
-}
-
-function RenderResults() {
-    const { results } = useMatches();
-
-    return (
-        <KBarResults
-            items={results}
-            onRender={({ item, active }) =>
-                typeof item === "string" ? (
-                    <div>{item}</div>
-                ) : (
-                    <div
-                        style={{
-                            background: active ? "#eee" : "transparent"
-                        }}
-                    >
-                        {item.name}
-                    </div>
-                )
-            }
-        />
     );
 }
